171214 22:24

+++ Body.h
@@ -19,15 +19,15 @@ public:
 	int intel;
 	int cx;
 	int cy;
-	float tiktak;		//Таймер... от него зависит то, сколько ты пропустишь.
-	float colldownS;	// скорость на колдунство.
-	float attackS;		// скорость физ атаки.
-	float fizCDS;		// скорость физ умений.
-	float moveS;		// скорость ходьбы.
-	float rundiagonalS; // скорость ходьбы на искосок.
-	float meditationS;  // скорость медитации восстановления маны и хп.
-	float shotS;		// выстрела скорость.
-	float jumpSpeed;	// скорость прыжка назад или вперед.
+	double tiktak;		//Таймер... от него зависит то, сколько ты пропустишь.
+	double colldownS;	// скорость на колдунство.
+	double attackS;		// скорость физ атаки.
+	double fizCDS;		// скорость физ умений.
+	double moveS;		// скорость ходьбы.
+	double rundiagonalS; // скорость ходьбы на искосок.
+	double meditationS;  // скорость медитации восстановления маны и хп.
+	double shotS;		// выстрела скорость.
+	double jumpSpeed;	// скорость прыжка назад или вперед.
 };
